Implement Lazy mode interface for Custom NMS OP

+++ tests/cpp/lazy_index_kernel_tests.cpp
@@ -182,7 +182,7 @@ TEST_F(LazyIndexKernelTest, NonZeroTestMixValues) {
       torch::randint(0, 7, {5, 7}, torch::dtype(torch::kInt64));
   torch::Tensor input_hpu = input_cpu.to(torch::kHABANA);
   auto out_hpu = torch::nonzero(input_hpu);
-  auto out_cpu = torch::nonzero(input_cpu);
+  auto out_cpu = torch::nonzero(input_cpu).to(torch::kInt32);
   auto h_cout = out_hpu.to(torch::kCPU);
   EXPECT_EQ(allclose(h_cout, out_cpu), true);
 }
@@ -192,7 +192,7 @@ TEST_F(LazyIndexKernelTest, NonZeroTestAllFalse) {
       torch::randint(0, 1, {5, 7}, torch::dtype(torch::kInt64));
   torch::Tensor input_hpu = input_cpu.to(torch::kHABANA);
   auto out_hpu = torch::nonzero(input_hpu);
-  auto out_cpu = torch::nonzero(input_cpu);
+  auto out_cpu = torch::nonzero(input_cpu).to(torch::kInt32);
   auto h_cout = out_hpu.to(torch::kCPU);
   EXPECT_EQ(allclose(h_cout, out_cpu), true);
 }
